login/requirements.txt
python-dotenv==1.0.0
python-jose==3.3.0
PyJWT==2.7.0
pydantic==1.8.2